+++ packages/bridge-status-controller/src/bridge-status-controller.test.ts
@@ -23,6 +23,7 @@ import type {
   MessengerEvents,
   MockAnyNamespace,
 } from '@metamask/messenger';
+import { CHAIN_IDS } from '@metamask/transaction-controller';
 import {
   TransactionType,
   TransactionStatus,
@@ -52,7 +53,6 @@ import type {
 import * as bridgeStatusUtils from './utils/bridge-status';
 import * as transactionUtils from './utils/transaction';
 import { flushPromises } from '../../../tests/helpers';
-import { CHAIN_IDS } from '../../bridge-controller/src/constants/chains';
 
 type AllBridgeStatusControllerActions =
   MessengerActions<BridgeStatusControllerMessenger>;